Push branch/initial push wizard
With certain workflows (non-Gerrit), creating a branch and pushing it to
a remote (including setup of branch tracking) is a common operation.
Currently, this is pretty complicated with EGit and a source of errors,
bug reports and help requests.

This change tries to provide a solution for the following use cases:

* The user created a new repository and wants to push it for the first
  time (probably one branch and no remotes yet).

* The user created a new feature/release branch and wants to push it to
  a remote, and set up tracking in the same step (git push -u).

It adds an additional wizard for both the above cases. For the toolbar
push action, the wizard is started when no upstream branch is configured
yet.

It's also available in the Team menu and the context menu of the
repositories view (on repo and on individual branches).

Bug: 382190
Change-Id: I7a84707e0dad78789075bdec97629c76f98c0dd7
Signed-off-by: Robin Stocker <robin@nibor.org>