Update patch set 8
Patch Set 8: Verified; Looks good to me, approved; IP review completed

- rebased
- fixed typo in commit message

timeout in this case means that execution of the Hudson was slow so that it exceeded the max build time configured for the job. Usually this happens if Hudson feels unhappy or if other heavy jobs are running concurrently.

Patch-set: 8
Label: SUBM=+1