Update patch set 2
Patch Set 2:

Because we're maybe going to have more options for merge (e.g. strategy options), I'd prefer to have a plain and fast "Merge" entry (like now), and an Advanced entry, which will open the dialog.

Maybe the advanced entry could be in a new Advanced submenu, as "Merge...". There could also be a "Rebase..." entry there in the future.

We have to take care that the menu will not get much bigger than it already is.

Patch-set: 2