Update patch set 2
Patch Set 2:

+1 for keeping the plain and fast "Merge" entry.

What about having the "Merge..." entry next to "Merge" when a (non checked out) branch is selected in the Repo view or a (non checked out) commit in the History View? It would open the wizard, but now with populated fields/selection.

When we put "Merge..." in an "Advanced" submenu in this case, but directly into the menu in all other cases, wouldn't that be inconsistent? Maybe not, but I'm not sure.

However, I would do that in an additional change, because with this change everything works fine. And the dialog for "Merge..." in the History View would be some effort, because we then would have to offer Commits instead of branches. Or alternatively, we could hide the branch/commit selection (because the user already has selected that) and just present the additional options?

Patch-set: 2