Improve hard wrapping by joining the line after a wrapped line
This extends the hard wrapping of the commit message in the Staging View
and Commit Dialog.

With the previous behavior, if one had a finished line/paragraph and
added text at the front, the text at the end would get
wrapped
like
this.

That made manual re-wrapping necessary. In the case of a paragraph,
several lines had to be edited again.

This change improves that use case by turning the line delimiter after a
wrapped word into a space to join the lines. It takes care not to join
empty lines or lines which start with a symbol (e.g. a bullet list).

It also gets rid of flickering during wrapping by only redrawing after
having applied all edits.

Bug: 387932
Change-Id: I197270a68700e7aacf12585c79418c7f8fc8a131
Signed-off-by: Robin Stocker <robin@nibor.org>