Update patch set 1
Patch Set 1:

You don't need to abandon a change in order to improve it. Next time better amend the commit you want to correct and push it again to refs/for/master. Since amend doesn't touch the Change-Id which was in the first patchset Gerrit will make your amended commit a new patchset for the already existing change. This way reviewers can use Gerrit's UI to compare the new patchset with previous ones.

Patch-set: 1