Add command to enable running garbage collector
The command is exposed in the context menu of repository nodes in the
Repositories View. If multiple repositories are selected they are gc'ed
one after the other.

JGit change I8dcdf0b83e91e6132dc490e8ec53818220773c94 isn't strictly
required for this change but fixes how the gc progress is displayed in
Eclipse.

Change-Id: Id5ef0b96a791187559efcdd032a0b929057cf851
Signed-off-by: Matthias Sohn <matthias.sohn@sap.com>