Update patch set 4
Patch Set 4: Verified; Looks good to me, but someone else must approve; IP review completed

Maybe we could extend this and provide an option which blocks the commit command if there are any problem markers on files to be committed. Though this needs to be optional as sometimes you may want to commit a draft in order to switch to another task.

Patch-set: 4