Update patch set 1
Patch Set 1:

Thanks. It apparantely did not achieve what it was supposed to do. 

Unfortunately isIgnoreHint seems pretty fast, so If noone can prove there is a performance benefit we should leave the code as is. The reporter thought this was a problem, but I don't see any significant performance boost from fixing bug 330917

Patch-set: 1