Add a statistics page to repository properties
When a user opens the properties of a repository (Git Repositories
View-> right-click on a repository node -> Properties) a new page is
added to the Component. This page "Statistics" contains info about the
number and size of loose and packed objects. Also the number of loose
an packed refs and the number of packfiles are listed.

Change-Id: I61281be7db512ccdf5399c98a90b39000365dd71