Ensure consistent ignore handling compared to cgit
EGit considers derived resources as if they were mentioned in
.gitignore. This looks like a nice Eclipse integration on first sight,
but it causes trouble when git is used on the command line (since
derived resources are not ignored there). This might also break
consistency when some developers use a different type of IDE or editor
which doesn't share Eclipse's opinion about which Resources are derived.
Fix this by no longer considering this hint provided by the Team
framework.

Bug: 359052
Change-Id: Ida2be440382b74715644e55fce9e8572ad0a8336
Signed-off-by: Matthias Sohn <matthias.sohn@sap.com>