Update patch set 3
Patch Set 3:

Ok, I tried that, but it seems like a general problem, not only for this change.

For me, when having the two resources selected, the history view still listed only one of the files in the title (in "File: ..."), instead of both. And it also just shows the history for that. Try the following:

* Select BaseDirCacheEditor.java
* Expand the selection to DirCache.java

-> It only shows one commit, the one for BaseDirCacheEditor.java. I also added a breakpoint to inputSet() of GitHistoryPage, and it is only called when the first resource of the selection changes.

On the other hand, when using Team > Show in History, it correctly lists the two resources in the title as "2 resources: ..." and shows the combined history.

Patch-set: 3