Update patch set 1
Patch Set 1:

With that we lose the use-case that the user knows a commit-id from somewhere and wants to replace the file content with that contained in that commit. Comparing would still work in that case with the menu entry Compare With > History..., (but I'm not sure how many users understand the history view well enough to achieve that).

Maybe we could take this change and just add a Replace With > History... instead of adding filter buttons to the selection dialog.

Patch-set: 1