Update patch set 1
Patch Set 1:

Hm..

Well, currently you'd have to scan the whole list to find the commit in that case, right? That is already pretty awkward. 

Can you give an example for that use case? I don't get why a user would have an ID of a commit that is not in the filtered history of the file and then wants to replace the file contents with that.

It would probably make more sense to use the History view for this:

* Show history (of whole repository) for the file
* Find the commit using the Find toolbar
* Right Click -> Compare with Working Tree
* Replace left side with contents from right side

How would a Replace With -> History... work?

Patch-set: 1