Update patch set 1
Patch Set 1: Verified

> Can you give an example for that use case? I don't get why
> a user would have an ID of a commit that is not in the
> filtered history of the file and then wants to replace the 
> file contents with that.

I have to admit that I never had this use case myself. It was more a theoretical thought. I agree that in 99% of the use-cases the filtering of the commits makes more sense, and in the other 1% users can use the history view. Therefore I think we should submit your change.

Thanks!

> How would a Replace With -> History... work?

We could add a new menu entry... On the other hand, the history view is complicated enough and I doubt that many users understand the mixture of menu entries, some operating on the commit and some on the resource that is the entry of the history view.

Patch-set: 1
Label: SUBM=+1