Git Tree Compare: Unify workbench and PathNode tree code
WorkbenchTreeContentProvider had a serious problem in that it only
descended into folders which existed in the workspace. The reason for
that was that it used getChildren of its base class
WorkbenchContentProvider.

For example, when the base (workspace) version had these files:

  a
  b/1

And compare version had these:

  a
  b/1
  c/1

The "c/1" file should be shown as "deleted" in the view. The problem was
that the "c" folder was not contained in the workspace and so the
content provider never included "c/1". It did include deleted files
whose parent folder was known, but not others.

This could maybe have been fixed by introducing artificial folder nodes
(implementing IWorkbenchAdapter) in such a situation (but only for the
special case above).

Instead, this change unifies the workbench with the PathNode content
provider code. Care was taken to keep the appearance by using
WorkbenchLabelProvider for container nodes and continuing to work with
IFile where possible. The tree construction was also simplified.

Bug: 390886
Change-Id: I726c41f7073820cf80030cc59feabb5b4b033e4d