Propagate project text label decoration up to working set
This introduced a different type for resource mappings so that the
configured project text decoration is used for working sets instead of
the default container/folder decoration. It also collects repository
name and branch info to allow those decorations for working sets but
only if all projects within the working set share the same repository.

Bug: 369969
Change-Id: I4aaea78104b5ab74f819e72c07126b34425b00d3
Signed-off-by: Gunnar Wagenknecht <gunnar@wagenknecht.org>