Update patch set 2
Patch Set 2:

> There is some discussion at bug 369969. One proposal was "[ ]" for different repos in a working set. There is also the case where different repos are in a working set but all pointing to the same branch. Should that be visualized too?

Ah yes, that would be very nice. I think we should do that with this change while we're at it.

I think for different repos with different branches, no decoration is better ("").

So, we have the following cases:

1. All from same repository and same branch:
   [repo branch ?status]
2. All from same repo, but not same branch:
   [repo *] or [repo ]
3. All from same branch, but not repo:
   [* branch] or [ branch]
4. All different -> empty

I'm not sure if "*" or "" is better for the case where the values differ.

Patch-set: 2