Update patch set 1
Patch Set 1:

Ok, we also need to guard against commitId being null in getContentIdentifier (was already a problem before). Apparently it is allowed to be null, otherwise this would make no sense:

	boolean exists() {
		return commitId != null;
	}

I'm not following what the problem is with members, it just passes on its own commitId to its children.

Patch-set: 1