Update patch set 2
Patch Set 2:

> 2. All from same repo, but not same branch:
>    [repo *] or [repo ]

Can this really happen? I thought that all projects from the same repo do always have the same branch?

Also note, the proposed implementation calculates branch name and status only once because it's expensive. If we want to show a '*' I have to do the calculation for every project until there is at least a difference. Worst case means that the calculation happens for all projects but they are all at the same repo and branch.

Patch-set: 2