Update patch set 2
Patch Set 2:

>> 2. All from same repo, but not same branch: > [repo *] or [repo ]

> Can this really happen? I thought that all projects from the same repo do always have the same branch?

Err, I meant when the repo name is the same (e.g. two different clones). It's probably rare but possible.

> Also note, the proposed implementation calculates branch name and status only once because it's expensive. If we want to show a '*' I have to do the calculation for every project until there is at least a difference. Worst case means that the calculation happens for all projects but they are all at the same repo and branch.

The most expensive to calculate is the branch status, as that needs to walk commits. Why not build a Set<Repository> first in the project loop, and then after the loop, do the following:

    if (repositories.size() == 1) {
        // Calculate repo name, branch name and status
    } else {
        // Loop over repositories
        // Stop calculating repo names if they are > 1 (using a set or some variables)
        // Do the same for branch names
    }

Branch status is not needed in the else, as that would be confusing. E.g. it would show ?1, but the actual number of commits to push would be > 1 because it's from multiple repos.

Patch-set: 2