RevisionSyntaxException in the Push wizard
The push dialog's refspec panel passed surplus text to the revision
parser. The revision parser used to allow garbage after the revision,
but was tightened, which then triggered this bug.

Remove the part following the branch name that is just UI decoration
before sending the branch name to the revision parser.

Bug: 393227
Change-Id: Ib21e0edfdd50687a67375435654de41140028266