Use three-way mode of the compare editor
When the user compares versions which are not direct descendants of
each other it is useful to see the content of the common ancestor.
This change introduces the usage of the three-way compare editor.

Three-way compare is still not supported when the user compares folders
or projects. This would have to be implemented in the CompareTreeView.

Change-Id: Ied1fa51d3a1ec9136ec270ff8500e08a32629fce
Signed-off-by: Stefan Lay <stefan.lay@sap.com>