Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

The code in ComboLabelingSupport is kind of strange. In setProposals, the items are set to the labels, in the selection listener of the constructor, it does setText with the content. So in the selection listeners, getText returns the content, but otherwise it returns the label as set by SWT.

How about doing the fix in ComboLabelingSupport? Add a getCurrentContent to it that does the following:

    String text = combo.getText();
    for (IContentProposal p : proposals)
        if (text.equals(p.getLabel()))
            return p.getContent()
    return text;

Then in the validateCreationPanel, use creationSrcComboSupport to get the src instead (same for dst). I would also replace the other usages of getText and remove the strange setText code in SelectionListener.

I can prepare a patch if you want.

Patch-set: 1