Fix RevisionSyntaxException in the Push wizard
The refspec panel passed strings like "master [branch]" to the revision
parser. The revision parser used to allow garbage after the revision,
but was tightened, which then triggered this bug.

Instead of working with the content proposal label, add a method to get
the corresponding content. Then use this when working with ref specs.

Another unhandled RevisionSyntaxException was in tryResolveLocalRef,
e.g. when calling completion with "refs/".

Bug: 393227
Change-Id: Id9b21ab851bb715264c304f08821df5323698d80