[stagingView] Restore previous commit message when amend is deselected
Sometimes the user clicks on amend by accident, which would cause the
carefully crafted commit message to be replaced by the amended message.

Now it is saved when selecting amend and restored when deselecting
again. When amend is selected and the user switches to another
repository and then back, the previous commit message can not be
restored as the save state has been overwritten. Oh well.

Bug: 364022
Change-Id: Icdf56dec24459aa66cc9bf8ab70aa957de39c31d