Update patch set 1
Patch Set 1:

Couldn't find such a case here, with jgit repo. When the file is not in the workspace (e.g. root pom.xml), the show in menu is expected to be empty, which works here.

Does failure depend on the file or does it fail/work some of the times for the same file?

Patch-set: 1