Update patch set 1
Patch Set 1:

> how can we ensure that old psf feature is deinstalled on upgrade ?

Hm, I don't know. The old psf feature lying around may indeed cause problems because it contains this in org.eclipse.egit.psf-feature/feature.xml (for 2.1):

    <requires>
          <import plugin="org.eclipse.egit.core" version="2.1.0" match="equivalent"/>
          ...

So when upgrading egit.core to 2.2, the dependency will no longer be satisfied because 2.1 is removed, right?

Maybe then we first need to change this dependency to greaterOrEqual in this cycle and then remove it in the next cycle.

Patch-set: 1