[repoView] Add Show In support for resources
The existing Show In commands for the repository node (History, Reflog,
Properties) are kept as is, as we don't want e.g. the Reflog view to
appear in the Show In menus everywhere.

So this only adds the platform Show In support for the existing
resources in the working dir tree nodes.

To continue to support showing multiple files in the history view, or
files outside of the workspace, we also pass a HistoryPageInput to the
ShowInContext if possible.

Also note that due to bug 383815 in platform UI, <visibleWhen> for
<dynamic> is broken in 4.2. To work around this, the new Show In is
defined in a separate <menu> so that <visibleWhen> can be moved there,
so it is evaluated for the menu. This does not result in two Show In
menus as the conditions don't overlap.

Bug: 358936
Change-Id: I7d986900e713b581b7baa19a8e850a5ee37585a7