[repoView] Add Show In support for resources
The existing Show In commands for the repository node (History, Reflog,
Properties) are kept as is, as we don't want e.g. the Reflog view to
appear in the Show In menus everywhere.

So this only adds the platform Show In support for the existing
resources in the working dir tree nodes.

To continue to support showing multiple files in the history view, or
files outside of the workspace, we also pass a HistoryPageInput to the
ShowInContext if possible.

Bug: 358936
Change-Id: I7d986900e713b581b7baa19a8e850a5ee37585a7