Compare using model provider if possible
The previous implementation of 'asCompareInput' hijacked any comparison
using model providers through the synchronize view. Even when
double-clicking on files that are part of a model, only the "git" model
would have been allowed to provide comparison input. This commit
modifies to implementation to prefer the model input if available.

Bug: 393225
Also-by: Gunnar Wagenknecht <gunnar@wagenknecht.org>
Change-Id: I69c3e36b74f8eee7755a55ec9f8f4cc42149fef7