Update patch set 1
Patch Set 1:

I'm investigating the test case failures. It seems that the comparison now does not honor the "local changes should be included in comparison" settings. This causes the the left side of the compare editor (in the test case) to always contain the "Locel File Workspace" leading to the test failure

Patch-set: 1