Update patch set 2
Patch Set 2: I would prefer that you didn't submit this

The commit message is still confusing. The current manifest in 'master' already (almost, see next paragraph) has the correct lower bounds for Juno, so this change here does not affect that - it just adds upper bounds.

Besides that, the lower version for team.ui should be 3.6.200 not 3.6.100.

Patch-set: 2