Update patch set 2
Patch Set 2: Do not submit

The correct approach is to get rid of the custom 'Show In' menu at all and use the built-in support, so that we do not get different menus when using the context menu and Navigate > Show In (Alt+Shift+W).

If this can't be fixed via 'plugin.xml' contribution, then we should simply contribute the menu like we do it for all the other views, i.e. in the code.

Patch-set: 2