[stagingView] Improve usability of "Replace with" on modified
The context menu actions "Replace with File in Git Index" and "Replace
with HEAD revision" in the unstaged changes table have the same result
for MODIFIED files. Showing both is redundant and can confuse newcomers.

Therefore remove the replace with index action, as the other is easier
to understand (the mental model of the user may be that the index only
contains changed files).

Both actions are still shown for PARTIALLY_MODIFIED, where they actually
cause different results.

Change-Id: I65601162f7ef506ea270933218601322fde1e929
Signed-off-by: Matthias Sohn <matthias.sohn@sap.com>