Update patch set 2
Patch Set 2:

Hm, ok. I'll experiment with IShowInTargetList to see if we can use that so that the "Git Reflog" entry only appears when the selection is a repository. If that doesn't work the other option would be to just show "Git Reflog" in all Show In menus, but I would prefer the first option.

Patch-set: 2