Use model compare input if possible
The implementation of 'asCompareInput' may hijacked any comparison
using model providers through the synchronize view. Even when
double-clicking on files that are part of a model, only the "git" model
would have been allowed to provide comparison input. This is a
limitation of the current synchronize model that happens when a file
from the local workspace is compared but without considering the working
copy contents (i.e. HEAD).

This commits works around the limitation by allowing the model compare
adapter to be used when possible, i.e. a comparison with the local
workspace is performed.

Bug: 393225
Also-by: Gunnar Wagenknecht <gunnar@wagenknecht.org>
Change-Id: I69c3e36b74f8eee7755a55ec9f8f4cc42149fef7