[repoView] Add Show In support for resources
Also convert other Show In commands (for repository, remote, fetch node,
push node, ref) to the platform Show In support.

To only list "Reflog" when a repository is selected, IShowInTargetList
is used.

To continue to support showing multiple files in the history view, or
files outside of the workspace, we also pass a HistoryPageInput to the
ShowInContext if possible.

Bug: 358936
Change-Id: I7d986900e713b581b7baa19a8e850a5ee37585a7