Enable Show In > Git Repositories for non-workspace files
By adding an IPath to the selection when the corresponding resource does
not exist in the workspace, we can show them at least in the Git
Repositories view (though not in Navigator).

For this to work in the Staging view, it depends on change
Ifdda37fc51a0cb29145e1c29b09d45317fc068ea so that StagingEntry#getFile()
returns null for non-workspace files.

Bug: 363567
Change-Id: Icf74705ac1b71d71b2052973c177e093f914cdba