Update patch set 1
Patch Set 1:

Filtered resources must also be ignored (e.g. for hierarchical project layouts).

I'm not sure how they work and if they need to be ignored explicitly. Just wondering if there is potential in designing a "should this resource be ignored" method in a way that it could also work for filtered resources. (Sorry that this is bit vague, I don't have access to the code right now.)

Patch-set: 1