Update patch set 4
Patch Set 4:

I've added a test class with a few test cases related to linked resources.
Not sure why the build on hudson failed. The test on which it failed passed fine here, but another test failed here but not on hudson.
Overall I don't get consistent results with swtbot, and as a result don't rely on it for submitting code.
Is there anything I should be aware to make these SWT tests work all the time?
I'm on eclipse 3.7.2 on linux 64bit.


Patch-set: 4