Update patch set 4
Patch Set 4:

Properties [Alt+Enter] must not be removed. This should be in all context menus as last item (currently not the case in the Repo view ==> a bug).

The next patch set should also use the new helper that creates the 'Show In' menu with the key binding in the label.

Patch-set: 4