Update patch set 4
Patch Set 4:

> Properties [Alt+Enter] must not be removed. This should be in all context menus as last item (currently not the case in the Repo view ==> a bug).

It's not removed, see the command item for org.eclipse.ui.file.properties. It's only available for repository nodes though. But this is a separate issue.

OpenPropertiesCommand was the manual (and badly-named) implementation of Show In > Properties.

> The next patch set should also use the new helper that creates the 'Show In' menu with the key binding in the label.

The key binding is already shown here because we use <menu commandId="org.eclipse.ui.navigate.showInQuickMenu" ... Does it not show the key binding for you?

Patch-set: 4