Update patch set 4
Patch Set 4:

I tested it with 3.7 as a target platform, the key binding is shown in the context menu there. With 4.2 as target platform, it doesn't work. So I guess it's a bug in 4.2.

I also tried egit-4.2-staging.target, but Eclipse doesn't start because org.eclipse.emf bundles are missing.

Patch-set: 4