Update patch set 4
Patch Set 4:

I filed bug 394958 for the platform issue.

The key binding works in 4.x in the other EGit views with the helper you wrote. All other views in the SDK (except for the marker views) also do it like this, so I suggest we do the same for the 'Git Repositories' view. In addition, using the same approach everywhere makes it easier to understand and later update the code.

Patch-set: 4