Update patch set 6
Patch Set 6:

Actually I just thought of some particular case that probably needs to be catered for: linked resources pointing to resources outside the project folder in the file system but still within the same git repository. In such case the linked resource can be operated on. With the recent change that makes sure we use IPath instead of IResource, maybe that should work.
Thoughts anyone?

Patch-set: 6