Update patch set 5
Patch Set 5:

it seems like adding files to the any of the known file-list will never treat the path the file is contained in in any of the directory-lists. i think the code would have to check for files whether their (or any parent) path is in any of the directory-lists and handle it accordingly. thoughts?

Patch-set: 5