Update patch set 4
Patch Set 4:

> Am I supposed to add a <actionProvider> to the org.eclipse.ui.navigator.navigatorContent?

Yes. Note that currently the content extension does not specify an enablement and/or possible children. This needs to be added in order to to add action providers. In RepositoriesViewActionProvider.fillContextMenu(IMenuManager) you would then add the 'Show In' menu as done in the other views (using your helper) and make sure it appears at the right location. One would normally use "group.show" (ICommonMenuConstants.GROUP_SHOW), but the Repo view context menu is pretty messed up, so I guess this would place it somewhere at the top.

Patch-set: 4