Update patch set 6
Patch Set 6:

Markus: I rebased this change onto a change which contains a test case for the problem. I suggest we now extend this change so that the test case passes. Do you want to work on that?

I think the solution needs two things:

In IndexDiffCacheEntry#calcIndexDiffData, check the untracked folders of the current indexDiffData if they contain one of the filesToUpdate paths. Add all those to the diffForChangedResources filter.

In the IndexDiffData constructor which merges the two, fix the treatment of untracked folders. I think all existing untracked folders containing one of the changed files need to be removed, and all the new untracked folders need to be added.

Patch-set: 6