Revert "Enable "Link with Editor" for compare editor from Tree Compare"
This reverts commit 7a406934a668abcf665edc5de9830f08b2065755 (change
I3c73212eadb883c598aee1ef8af882e0f7725899).

It was causing multiple regressions:

When modifying the contents in the compare editor, it could not be
saved, see bug 393990.

When using the Merge Tool with the "Use HEAD" option, the contents of
the workspace file were shown instead of that of the passed file
revision, see bug 395177.

Bug: 393990
Bug: 395177
Change-Id: I05374ccab215a320f0c808cae9d22f99a7fd5399