Gerrit doesn't need /p anymore
Since Gerrit 2.3 the path segment /p is not needed anymore when
accessing Gerrit with the http protocol [1]. This patch was initially
created because there was a bug in Gerrit when pushing with /p [2],
which is fixed in the meantime. When testing the change, bug 407984 was
detected. Too fix bug 407984, avoid unnecessary complexity and because
Gerrit 2.3 is now out for half a year, the logic to add /p is removed
from the Gerrit Configuration wizard.

[1]
http://gerrit-documentation.googlecode.com/svn/ReleaseNotes/ReleaseNotes-2.3.html#_miscellaneous
(Support http://server/project for Git access)
[2] https://bugs.eclipse.org/bugs/show_bug.cgi?id=393735
 
Bug: 407984
Change-Id: I4bcadeb323220918c13e9a43205d9546ad75096e
Signed-off-by: Matthias Sohn <matthias.sohn@sap.com>
Signed-off-by: Stefan Lay <stefan.lay@sap.com>