Simplify Eclipse setup for EGit developers
This provides a quick way to install the tools typically used by
JGit/EGit contributors. In addition it provides an alternative to using
a target platform for provisioning 3rd party dependencies required to
compile and test JGit and EGit. Using a target platform is more accurate
as it allows to separate the dependencies of the workbench used for
development from the one used for testing. But using a target platform
is typically slower since PDE tries to communicate with p2 repositories
listed in the target platform.

In order to install dependencies and plugins typically used by EGit
developers click "Import... > Install > Install Software Items from
File" and select this p2f file, then select the bundles you want to
install and click OK.

This works for Eclipse Juno SR1 and Kepler M3.

Change-Id: I84709b302297e8ad9fb8c1b11d2edb85c7f03c57
Signed-off-by: Matthias Sohn <matthias.sohn@sap.com>