Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

You forgot to link the bug: 393735

This works well, but I would prefer the Gerrit bug to be fixed, in order not to loose compatibility with older versions, as Dani had already stated. It seems that pushing without /p and the http protocol only works due to another issue.

Sasa, what do you think? Since which Gerrit version does pushing over http work without /p, and since which version it does not work with /p?

When testing, I found a bug in the old implementation: When the path to the repo started with /p (e.g. /pub for the linux kernel), switching to http did nothing, but switching back removed the first segment. If we keep the /p rule for http, I will fix that.

Patch-set: 1