[sync] Update tracking branch when pushing
Pushing a branch to a remote for which there is a tracking branch should
update the tracking branch. This was not the case when using the "Push"
action in the Synchronize view.

The solution involves two parts. First, PushOperation did not consider
fetch refspecs while calculating RemoteRefUpdates. Changing that caused
the RemoteRefUpdates to correctly include a TrackingRefUpdate.

Second, PushOperationUI used to take the passed RemoteRefUpdates and
then create RefSpecs for passing to the PushCommand. In this conversion,
tracking ref updates are lost. By using the transport API we can
directly pass the RemoteRefUpdates.

Another solution would have been to pass the remote name to the
PushCommand instead of a bare URIsh, but that would have involved
changing PushOperationSpecification (which only uses URIish).

Bug: 317411
Change-Id: I0e43f8b4e0b525a5cf35de9fdc135e2e539d03b6