Can not delete git project
Bug: 395488
Steps to reproduce:
- ensure the Git decorator is enabled
- import a project to Eclipse using the Import Project from Git wizard
- try to delete the project including its contents
Deleting will fail
- delete the project without its contents
- try to delete the project's directory using an OS command
Deleting will fail
The issue can always be reproduced on Windows. As to Linux, it can be
reproduced when cloning the project to a mounted NFS device.
For more details, see https://issues.jboss.org/browse/JBIDE-10398.
Change-Id: Iae1be7fa4c3419f2a7fddcacb04679e85c83d490
Signed-off-by: Snjezana Peco <snjeza.peco@gmail.com>
I have authored 100% of the content I'm contributing
I have the rights to donate the content to Eclipse