Update patch set 5
Patch Set 5:

Changed the context menu to be provided by an action provider. I don't know if changing <triggerPoints> to <enablement> is ok there.

And yes, the Show In now shows up at the top. I'm not sure that's bad, the Package Explorer also shows it in the top part. The right way to fix this would be to add <popupMenu>  and <insertionPoint> to the viewer extension and then separately attach menu contributions to the groups, right?

Patch-set: 5