Make RemoveFromIndexOperation work for multiple repositories
RemoveFromIndexActionHandler is enabled when resources from different
repositories are selected. But before this change it would not actually
work because it passed the repository to the operation and the operation
would use that to calculate the paths.

The API is also changed to work directly with IPath, so that in future
the operation can also be used for non-workspace files (e.g. from the
Repositories view).

Bug: 395722
Change-Id: I26c3a84aa8e874d74fe863428cb737031757b372