Support filtered resources in ContainerTreeIterator
Filtered resources are useful when working with nested projects, like
this:

  root/
    .project
    sub/
      .project

In the above configuration, the root project typically contains resource
filters so that "sub" is not shown as a subdirectory of root in
navigator views but can be imported as a separate project.

When resources are filtered, they are not returned from API like
IContainer#members. So ContainerTreeIterator needs to check if resource
filters are configured and if that's the case, list additional
files/folders through the Java File API. It then falls back to
AdaptableFileTreeIterator for subtrees.

This depends on small changes in JGit's FileTreeIterator, see below.

Bug: 358901
Change-Id: I2c41d86b8b74c2a334433de1bbfed5b36af872bf
JGit-Dependency: Ie0f9c9434ef7d73a8d73d4fe46db4147ded1d267