Update patch set 9
Patch Set 9: I would prefer that you didn't submit this

I didn't look at all the changes in detail but there are quite a lot of them hence I would prefer not to submit this just 2 weeks before we ship 2.2. So, the -1 is only for the upcoming 2.2 and not a general one.

Patch-set: 9