Update patch set 10
Patch Set 10:

I've reduced the scope of this change in the latest patch set as recommended in previous comments. It's still a big change, so I understand the desire to merge it after egit 2.2. Just bear in mind however that I've been keeping it up-to-date since June last year. Please let's make it a point to review and merge it just after egit 2.2 gets released.

Patch-set: 10