Update patch set 5
Patch Set 5: I would prefer that you didn't submit this

It looks like the failure in GitRepositoriesViewTest/testShowProperties is caused by the changes.

For the context menu on the repository node it looks a bit ugly when 'Show In' is at the top. I think we should try to put it at the same location as it is now. I guess this can be done the way you mentioned. See also bug 388208.

Otherwise the changes look OK to me.

Patch-set: 5