Update patch set 3
Patch Set 3: I would prefer that you didn't submit this; Unclean IP, do not check in

(3 inline comments)

Please follow [1] and mark draft changes accordingly using [DRAFT] or [RFC] in commit message header to avoid wasting time of those reviewers which are not interested in incomplete work.

Does your last name really end with "?" ?

We need you to assert the contributor questions as a comment on every change you upload to Gerrit [2].

In general we strive for good test coverage. So if you add a feature or fix a bug also add or enhance tests.

[1] http://wiki.eclipse.org/EGit/Contributor_Guide#Submission_Guidelines
[2] http://wiki.eclipse.org/EGit/Contributor_Guide#Legal_Paperwork

Patch-set: 3