Update patch set 1
Patch Set 1:

It's not about the push specs, it's about the third argument to findRemoteRefUpdatesFor, which is used for this:

	 * @param fetchSpecs
	 *            fetch specifications used for finding localtracking refs. May
	 *            be null or empty collection.

They are only used for calculating the tracking refs to update, i.e. in a typical configuration the fetch refspec is defined like this:

[remote "origin"]
        fetch = +refs/heads/*:refs/remotes/origin/*

Which results in e.g. refs/remotes/origin/master being updated when pushing to refs/heads/master.

Or maybe I misunderstood your question?

Patch-set: 1