Update patch set 1
Patch Set 1:

With this change, tracking refs will only be updated when the right part of a push spec matches the left part of a fetch spec.

In your example, e.g. when the local refs/heads/master is pushed to refs/remotes/origin/q/master on the remote (because of the push spec), transport checks if refs/remotes/origin/q/master is contained in the left part of the passed fetch spec. In this case, it isn't, so no tracking ref is updated.

The push spec in the example is a bit strange as the right part is a funny ref to push to on the remote. Another example would be a push spec of "refs/heads/*:refs/heads/q/*". So when pushing "master" to "q/master", the tracking ref "origin/q/master" would be updated, which is correct.

Basically it should behave the same way as if a fetch was executed right afterward the push.

Patch-set: 1