Update patch set 4
Patch Set 4: Looks good to me, but someone else must approve

This is a simple move method refactoring. If it can be reused elsewhere (and we can see Gunnar has other patches depending on this) then it's worth it. +1 from me.

Let's wait for bug 396884 and double check with Hudson. To be clear, it's a formality.

Patch-set: 4