Update patch set 3
Patch Set 3:

This significantly simplifies the initial contributor setup, so I'm +1 on this. (Target platform is ok, but it's not enough  because e.g. execution environment descriptions and SWTBot still have to be installed manually.)

EGit and JGit are disabled here with the following text:

(Disabled because a higher version is already installed)

Anyone else seeing this?

Patch-set: 3