[repoView] Use <insertionPoint> for popup menu groups
This allows us to add a standard "group.show" insertion point which is
used in I7d986900e713b581b7baa19a8e850a5ee37585a7 to add the standard
Show In menu.

It also moves "Properties" to the bottom of the menu as this is the
common location.

"Open in Editor" is moved above "Show In", which corresponds to the
layout used in Package Explorer.

Bug: 358936
Change-Id: I9471738cbd8d3c3f968ae4fb3392f203a17471b5