Fix NPE in IndexDiffCacheEntry.calcIndexDiffData (incremental)
When an incremental refresh is done before any full refresh,
indexDiffData is null, which could lead to an NPE. So do a full refresh
instead when no prior indexDiffData is available.

This can happen because full reload jobs cancel each other, whereas an
incremental reload job is never cancelled. Whether this is desired
should be investigated separately.

This change also renames the methods for more clarity.

Bug: 396940
Change-Id: Icbacef742d498a91d579db41c27ae88d6a523729