RFC Make the ContainerTreeIterator recognize symbolic links
We cannot rely on Eclipse's resource handling here since it
looks at the link target rather than the link itself.

This depends on the support for symbolic links in JGit. I will
only work when Eclipse runs with Java7 on a platform that
suports symbolic links like unix and possibly newer versions
of Windows.

Change-Id: I94bfd493805478f2a232229ce6d25a0d810bfd0f