Show untracked files if required by preferences
The "show untracked" dialog setting in the commit dialog is not
synchronized with the "include untracked" preference option. This is
wrong because both are related. If the preference setting is true then
untracked files are checked but they may not be visible in the commit
dialog. 

This commits links the enabled states together. As a result, when
the preference is set to true, the action "show untracked" will always
be enabled when the commit dialog pops up.

Change-Id: I295b667469420a48299fd454f402c416e17aba27
Signed-off-by: Gunnar Wagenknecht <gunnar@wagenknecht.org>