Update patch set 3
Patch Set 3:

Though, not a reason for rejection since this is a new feature and I was annoyed before this. 

Filtering is very very slow when the number of entries is huge, say tens of thousands which would be case when adding entirely new projects. There are other immensely slow things (patches ahead) there though. I'll probably post some patches there.

Is it possible to address this? Among the unnecessary things I observed is that besides pattern matching, PatternMatcher breaks the filename into "words" and performs a literal match on each part, which seems broken.

Patch-set: 3