Improve precision in detacted head decoration
For decorations in EGit we want to display the name used for checking
out a tag. HEAD point to a commit id directly, but the reflog contains
the name used. Scan the reflog and use the latest checkout whose refname
resolves to the same commit as HEAD.

If not try, as before, try to find any ref and use the one that matches
prioritized by latest update.

Bug: 385392
Change-Id: Icb10545a1e503a98fd9ea98fbb6770c37a9e4bfe