Update patch set 1
Patch Set 1:

https://hudson.eclipse.org/sandbox/job/egit.gerrit/3956/testReport/junit/org.eclipse.egit.ui.view.repositories/GitRepositoriesViewFetchAndPushTest/testPushToOriginPushNode/ isn't caused by https://git.eclipse.org/r/#/c/9564/

The failure happens randomly (as it is in https://hudson.eclipse.org/sandbox/job/egit.gerrit/3950/, for instance).
I think you need to wait for Eclipse jobs to finish before and after creating/deleting a project in the tests.
We have had a lot of issues related to the creating/deleting projects in tests that we have solved using the following utility class:
https://github.com/jbosstools/jbosstools-base/blob/master/tests/plugins/org.jboss.tools.tests/src/org/jboss/tools/test/util/JobUtils.java

Patch-set: 1