Update patch set 3
Patch Set 3:

Snjezana gave the hint [1] that the problem might be caused by jobs running concurrently with the tests and pointed out a solution they use for jboss tooling. Instead of their solution (reuse would require a CQ and is based on polling) we could try to delete test projects using a job with a workspace lock to ensure that no other jobs are running. I'll discuss this idea with Jens who is an expert in Eclipse job framework.

[1] https://git.eclipse.org/r/#/c/9564/

Patch-set: 3