Update patch set 19
Patch Set 19:

Rebased.

I tested the change a bit with a linked folder named root in org.eclipse.egit pointing to .. (the root of the egit repository working directory) and with resource filters for "org.eclipse.egit*".

This means that only the files such as README and the tools folder is visible.

I noticed a few things:

* The files from the tools folder is marked as deleted in the Staging view. Changing IteratorService to use isLinked(IResource.CHECK_ANCESTORS) fixes that (it used isLinked() before). Do you agree with this change?

* When editing README and saving, the change does not automatically show up in the Staging view. This is caused by GitResourceDeltaVisitor skipping the resource. When removing this, it works. What is the reason behind this? I didn't notice anything being wrong with the lines removed.

Patch-set: 19