Improve label of index version in compare editors
When opening the compare editor from the staging view for an unstaged
file, the right side shows the Index version.

The problem is that not many people discover that it is in fact editable
(and the way to do partial staging).

This change tries to help a bit by changing the label from

  file.txt Index ()

to

  Index: file.txt (editable)

Index is used as a prefix because the left side is "Local: file.txt".

Bug: 398562
Change-Id: I9940d2f148393b1f4704664e379c5c863565a118