Add Compare with ours/theirs actions for conflicts
The action's labels change depending on whether it's a merge,
cherry-pick or rebase conflict to provide better labels than the generic
"ours" and "theirs" (which is confusing especially during rebase).

The actions are only available in the Staging view yet, see bug 347835
for the problems with the Compare with/Replace with menus.

Bug: 391856
Change-Id: I6587c6c195bd61800fd32f46465cf1237efa6d31