Update patch set 24
Patch Set 24:

The test that failed runs fine on my machine, but there are errors in the console view:
!ENTRY org.eclipse.osgi 2 1 2013-01-27 22:08:13.347
!MESSAGE NLS unused message: DiscardChangesOperation_repoNotFound in: org.eclipse.egit.core.coretext

!ENTRY org.eclipse.equinox.event 4 0 2013-01-27 22:08:33.206
!MESSAGE Exception while dispatching event org.osgi.service.event.Event [topic=org/eclipse/e4/ui/model/ui/UIElement/visible/SET] to handler org.eclipse.e4.ui.services.internal.events.UIEventHandler@19b271ad
!STACK 0
java.lang.IllegalArgumentException: Argument not valid
	at org.eclipse.swt.SWT.error(SWT.java:4342)
	at org.eclipse.swt.SWT.error(SWT.java:4276)
and a bunch of non-egit stack trace, so it's more of a issue with the test I guess.

Anyway, I removed the unused text properties, so a new patch set is coming.

Patch-set: 24