Update patch set 4
Patch Set 4: Looks good to me, but someone else must approve

(1 inline comment)

Robin, that's a flaw in the current design. When the preference is set, untracked files will be included in the commit. Thus, the dialog must make sure that the committer recognizes this. If the dialog doesn't show them they might be accidentally committed without recognizing. I don't think that's the intent.

Patch-set: 4