Update patch set 1
Patch Set 1:

The problem is in GitResourceVariantTree, where getFullPath is called on an IResource which was previously obtained via IWorkspaceRoot#getFile. The problem is already with getFile, which expects a workspace path (/projectname/filename). Try this test case:

	@Test
	public void testFullPathWithDevice() {
		Path path = new Path("C:", "/workspaces/repo/file.txt");
		IWorkspaceRoot root = ResourcesPlugin.getWorkspace().getRoot();
		IFile file = root.getFile(path);
		assertEquals(path, file.getFullPath());
	}

It fails with:

	java.lang.AssertionError: expected:<C:/workspaces/repo/file.txt> but was:</workspaces/repo/file.txt>

So the device part of the path is lost in this roundtrip, and the result of stripWorkDir with this path will be a string like "le.txt" (because too much is stripped).

A pretty hacky way to deal with this here would be to detect the missing device part and handle it differently in stripWorkDir.

But I can imagine that the whole thing would break horribly when there is a project which is named "workspaces".

Patch-set: 1