GitFileHistory should return the full history of its target.
When using GitFileHistoryProvider.getFileHistoryFor(IResource,...), a
user expects the full history of the target resource to be returned,
except if the SINGLE_REVISION or SINGLE_LINE_OF_DESCENT flags are
positionned.

The SINGLE_REVISION flag is properly handled, however when no flags are
set, EGit previously returned only the commits descending from HEAD,
without considering the history that exists on other branches or
references.

This adds specific tests to make sure of the behavior when either no
flag or the SINGLE_REVISION flag are set.

The SINGLE_LINE_OF_DESCENT flag remains to be handled.

Bug: 398982
Change-Id: Id596d4501cdca815e9563a1f7f4a1d58bc3920c9