GitFileHistory should return the full history of its target.
When using GitFileHistoryProvider.getFileHistoryFor(IResource,...), a
user expects the full history of the target resource to be returned,
except if the SINGLE_REVISION or SINGLE_LINE_OF_DESCENT flags are
positioned.

The SINGLE_REVISION flag is properly handled, however when no flags are
set, EGit previously returned only the commits descending from HEAD,
without considering the history that exists on other branches or
references.

This adds specific tests to make sure of the behavior when either no
flag or the SINGLE_REVISION flag are set.

The history's "getTarget" and "getContributors" methods are tested to
return respectively all direct children or all direct parents of the
given commit.
Both now support some of Team's quirks, such as the synchronize view
returning its own implementations of IFileRevisions instead of the
supplied CommitFileRevisions. This required opening GitRemoteResource as
API (previously package-protected) so that we could adapt the file
revisions into a format known by EGit.

DRAFT: the current implementation exhibits a lot of issues when
exerted through pure team APIs. A specific test aimed at that is in a
failing, draft status.

Bug: 398982
Change-Id: Id596d4501cdca815e9563a1f7f4a1d58bc3920c9